Exclude OPTIONS from always supported HTTP methods
Add HEAD as an allowed method for web console e2e tests
